Check to see if song exists before adding it to database
There is now an async function that guards against inserting the same thing into the database. It looks for the song in in the database and if it already exists, it won't try to add it. Dexie won't overwrite info,  but it will throw an error if it tries to add the same info twice.