fix: Album play count no longer displays duplicate data points
The album view on the play count data section would group together data points by day. This meant
that two records in the same day would get stacked on top of eachother and it would create a day
with two entry points, with a small vertical line connecting them. This was less visually appealing,
so the first data point on duplicate entries are removed, showing only the point with the higher
play count.